Extracted Crummy::StandardRenderer
