+++ .gitignore
@@ -1,2 +1,3 @@
 rdoc
 pkg
+nbproject
